requirements.txt
Flask
gunicorn
requests
pillow
lxml
google-api-python-client
httplib2
oauth2client
flask-cors